+++ src/test/java/net/snowflake/client/jdbc/HeartbeatIT.java
@@ -3,15 +3,9 @@
  */
 package net.snowflake.client.jdbc;
 
-import net.snowflake.client.AbstractDriverIT;
-import net.snowflake.client.ConditionalIgnoreRule;
-import net.snowflake.client.RunningOnGithubAction;
-import net.snowflake.client.category.TestCategoryOthers;
-import net.snowflake.client.core.QueryStatus;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
+import static org.hamcrest.CoreMatchers.equalTo;
+import static org.hamcrest.CoreMatchers.instanceOf;
+import static org.junit.Assert.*;
 
 import java.sql.*;
 import java.util.ArrayList;
@@ -22,10 +16,14 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.logging.Logger;
-
-import static org.hamcrest.CoreMatchers.equalTo;
-import static org.hamcrest.CoreMatchers.instanceOf;
-import static org.junit.Assert.*;
+import net.snowflake.client.AbstractDriverIT;
+import net.snowflake.client.ConditionalIgnoreRule;
+import net.snowflake.client.RunningOnGithubAction;
+import net.snowflake.client.category.TestCategoryOthers;
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
+import org.junit.experimental.categories.Category;
 
 /** This test assumes that GS has been set up */
 @Category(TestCategoryOthers.class)
@@ -80,7 +78,7 @@ public class HeartbeatIT extends AbstractDriverIT {
    * @param queryIdx The query index
    * @throws SQLException Will be thrown if any of the driver calls fail
    */
-  private void submitQuery(boolean useKeepAliveSession, int queryIdx)
+  void submitQuery(boolean useKeepAliveSession, int queryIdx)
       throws SQLException, InterruptedException {
     Connection connection = null;
     Statement statement = null;
@@ -114,51 +112,12 @@ public class HeartbeatIT extends AbstractDriverIT {
   }
 
   /**
-   * create a new connection with or without keep alive sessionsubmit an asynchronous query that
-   * will take longer than the master token validity.
-   *
-   * @param useKeepAliveSession Enables/disables client session keep alive
-   * @param queryIdx The query index
-   * @throws SQLException Will be thrown if any of the driver calls fail
-   */
-  private void submitAsyncQuery(boolean useKeepAliveSession, int queryIdx)
-      throws SQLException, InterruptedException {
-    Connection connection = null;
-    ResultSet resultSet = null;
-    try {
-      Properties sessionParams = new Properties();
-      sessionParams.put(
-          "CLIENT_SESSION_KEEP_ALIVE",
-          useKeepAliveSession ? Boolean.TRUE.toString() : Boolean.FALSE.toString());
-
-      connection = getConnection(sessionParams);
-
-      Statement stmt = connection.createStatement();
-      // Query will take 30 seconds to run, but ResultSet will be returned immediately
-      resultSet =
-          stmt.unwrap(SnowflakeStatement.class)
-              .executeAsyncQuery("SELECT count(*) FROM TABLE(generator(timeLimit => 30))");
-      Thread.sleep(61000); // sleep 61 seconds to await original session expiration time
-      QueryStatus qs = resultSet.unwrap(SnowflakeResultSet.class).getStatus();
-      // Ensure query succeeded
-      assertEquals(QueryStatus.SUCCESS, qs);
-
-      // assert we get 1 row
-      assertTrue(resultSet.next());
-      assertFalse(resultSet.next());
-
-      logger.fine("Query " + queryIdx + " passed ");
-    } finally {
-      resultSet.close();
-      connection.close();
-    }
-  }
-
-  /**
    * Test heartbeat by starting 10 threads. Each get a connection and wait for a time longer than
    * master token validity and issue a query to make sure the query succeeds.
    */
-  private void testSuccess(boolean isAsync) throws Exception {
+  @Test
+  @ConditionalIgnoreRule.ConditionalIgnore(condition = RunningOnGithubAction.class)
+  public void testSuccess() throws Exception {
     int concurrency = 10;
     ExecutorService executorService = Executors.newFixedThreadPool(10);
     List<Future<?>> futures = new ArrayList<>();
@@ -171,11 +130,7 @@ public class HeartbeatIT extends AbstractDriverIT {
           executorService.submit(
               () -> {
                 try {
-                  if (isAsync) {
-                    submitAsyncQuery(true, queryIdx);
-                  } else {
-                    submitQuery(true, queryIdx);
-                  }
+                  submitQuery(true, queryIdx);
                 } catch (SQLException | InterruptedException e) {
                   throw new IllegalStateException("task interrupted", e);
                 }
@@ -189,19 +144,16 @@ public class HeartbeatIT extends AbstractDriverIT {
    * Test no heartbeat by starting 1 thread. It gets a connection and wait for a time longer than
    * master token validity and issue a query to make sure the query fails.
    */
-  private void testFailure(boolean isAsync) throws Exception {
+  @Test
+  @ConditionalIgnoreRule.ConditionalIgnore(condition = RunningOnGithubAction.class)
+  public void testFailure() throws Exception {
     ExecutorService executorService = Executors.newFixedThreadPool(1);
-
     try {
       Future<?> future =
           executorService.submit(
               () -> {
                 try {
-                  if (isAsync) {
-                    submitAsyncQuery(false, 0);
-                  } else {
-                    submitQuery(false, 0);
-                  }
+                  submitQuery(false, 0);
                 } catch (SQLException e) {
                   throw new RuntimeSQLException("SQLException", e);
                 } catch (InterruptedException e) {
@@ -220,30 +172,6 @@ public class HeartbeatIT extends AbstractDriverIT {
       assertThat("Root cause class", rootCause, instanceOf(SnowflakeSQLException.class));
       assertThat("Error code", ((SnowflakeSQLException) rootCause).getErrorCode(), equalTo(390114));
     }
-  }
-
-  @Test
-  @ConditionalIgnoreRule.ConditionalIgnore(condition = RunningOnGithubAction.class)
-  public void testSynchronousQuerySuccess() throws Exception {
-    testSuccess(false);
-  }
-
-  @Test
-  @ConditionalIgnoreRule.ConditionalIgnore(condition = RunningOnGithubAction.class)
-  public void testAsynchronousQuerySuccess() throws Exception {
-    testSuccess(true);
-  }
-
-  @Test
-  @ConditionalIgnoreRule.ConditionalIgnore(condition = RunningOnGithubAction.class)
-  public void testSynchronousQueryFailure() throws Exception {
-    testFailure(false);
-  }
-
-  @Test
-  @ConditionalIgnoreRule.ConditionalIgnore(condition = RunningOnGithubAction.class)
-  public void testAsynchronousQueryFailure() throws Exception {
-    testFailure(true);
   }
 
   class RuntimeSQLException extends RuntimeException {
